@@ -257,14 +257,14 @@ class Barcode {
     _data = _convertData(barcodeData);
   }
 
-  BarcodeType _type;
-  List<int> _data;
+  BarcodeType? _type;
+  List<int>? _data;
 
   List<int> _convertData(List<dynamic> list) =>
       list.map((dynamic d) => d.toString().codeUnitAt(0)).toList();
 
   int _charcode(dynamic ch) => ch.toString().codeUnitAt(0);
 
-  BarcodeType get type => _type;
-  List<int> get data => _data;
+  BarcodeType? get type => _type;
+  List<int>? get data => _data;
 }
