@@ -43,15 +43,15 @@ class CapabilityProfile {
   String name;
   List<CodePage> codePages;
 
-  int getCodePageId(String codePage) {
+  int getCodePageId(String? codePage) {
     if (codePages == null) {
       throw Exception("The CapabilityProfile isn't initialized");
     }
 
     return codePages
         .firstWhere((cp) => cp.name == codePage,
-            orElse: () => throw Exception(
-                "Code Page '$codePage' isn't defined for this profile"))
+            orElse: (() => throw Exception(
+                "Code Page '$codePage' isn't defined for this profile")) as CodePage Function()?)
         .id;
   }
 
